requirements.txt
Flask-Login~=0.6.3
Werkzeug~=3.0.6
Flask~=3.0.3
python-dotenv~=1.0.1
flask-migrate~=4.0.5
Pillow~=11.0.0
SQLAlchemy~=2.0.36
